+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
----
+## [1.0.0] Unreleased
+
+- Support for `delphi-logger` added (opt-in structural logging for debug
+purposes.) [#19](https://github.com/continuous-delphi/delphi-msbuild/issues/19)
 
 ## [0.7.0] Unreleased